pull out resource watching logic to _ServiceWatcher
