+++ src/hassette/core/core.py
@@ -17,10 +17,11 @@ from .apps.app_handler import _AppHandler
 from .bus.bus import Bus, _Bus
 from .classes import Resource, Service
 from .enums import ResourceRole, ResourceStatus
-from .events import Event, HassetteServiceEvent
+from .events import Event
 from .file_watcher import _FileWatcher
 from .health_service import _HealthService
 from .scheduler.scheduler import Scheduler, _Scheduler
+from .service_watcher import _ServiceWatcher
 from .websocket import _Websocket
 
 P = ParamSpec("P")
@@ -68,6 +69,7 @@ class Hassette:
         self._thread_pool = ThreadPoolExecutor(max_workers=10)
 
         # internal only (so far, at least)
+        self._service_watcher = self._register_resource(_ServiceWatcher)
         self._websocket = self._register_resource(_Websocket)
         self._app_handler = self._register_resource(_AppHandler)
         self._file_watcher = self._register_resource(_FileWatcher)
@@ -273,85 +275,10 @@ class Hassette:
 
         return True
 
-    async def restart_service(self, event: HassetteServiceEvent) -> None:
-        """Start a service from a service event."""
-        data = event.payload.data
-        name = data.resource_name
-        role = data.role
-
-        try:
-            if name is None:
-                self.logger.warning("No %s specified to start, skipping", role)
-                return
-
-            self.logger.info("%s '%s' is being restarted after '%s'", role, name, event.payload.event_type)
-
-            self.logger.info("Starting %s '%s'", role, name)
-            service = self._resources.get(name)
-            if service is None:
-                self.logger.warning("No %s found for '%s', skipping start", role, name)
-                return
-
-            service.cancel()
-            service.start()
-
-        except Exception as e:
-            self.logger.error("Failed to restart %s '%s': %s", role, name, e)
-            raise
-
-    async def log_service_event(self, event: HassetteServiceEvent) -> None:
-        """Log the startup of a service."""
-
-        name = event.payload.data.resource_name
-        role = event.payload.data.role
-
-        if name is None:
-            self.logger.warning("No resource specified for startup, cannot log")
-            return
-
-        status, previous_status = event.payload.data.status, event.payload.data.previous_status
-
-        if status == previous_status:
-            self.logger.debug("%s '%s' status unchanged at '%s', not logging", role, name, status)
-            return
-
-        try:
-            self.logger.info(
-                "%s '%s' transitioned to status '%s' from '%s'",
-                role,
-                name,
-                event.payload.data.status,
-                event.payload.data.previous_status,
-            )
-
-        except Exception as e:
-            self.logger.error("Failed to log %s startup for '%s': %s", role, name, e)
-            raise
-
-    async def shutdown_if_crashed(self, event: HassetteServiceEvent) -> None:
-        """Shutdown the Hassette instance if a service has crashed."""
-        data = event.payload.data
-        name = data.resource_name
-        role = data.role
-
-        try:
-            self.logger.exception(
-                "%s '%s' has crashed (event_id %d), shutting down Hassette, %s",
-                role,
-                name,
-                data.event_id,
-                data.exception_traceback,
-            )
-            self.shutdown()
-        except Exception:
-            self.logger.error("Failed to handle %s crash for '%s': %s", role, name)
-            raise
-
     async def run_forever(self) -> None:
         """Start Hassette and run until shutdown signal is received."""
         self._loop = asyncio.get_running_loop()
         self._start_resources()
-        self._register_internal_event_listeners()
 
         self.ready_event.set()
 
@@ -393,12 +320,6 @@ class Hassette:
                 await resource.shutdown()
             except Exception as e:
                 self.logger.error("Failed to shutdown resource '%s': %s", resource.class_name, e)
-
-    def _register_internal_event_listeners(self) -> None:
-        """Register internal event listeners for resource lifecycle."""
-        self.bus.on_hassette_service_failed(handler=self.restart_service)
-        self.bus.on_hassette_service_crashed(handler=self.shutdown_if_crashed)
-        self.bus.on_hassette_service_status(handler=self.log_service_event)
 
     async def _shutdown(self) -> None:
         """Shutdown all services gracefully and gather any results."""
